domain_client_operator::domain_bundle_producer

Type Alias BundleHeaderV0For

pub type BundleHeaderV0For<Block, CBlock> = BundleHeaderV0<NumberFor<CBlock>, BlockHashFor<CBlock>, HeaderFor<Block>, Balance>;

Type alias for bundle header V0.

Aliased Type

pub struct BundleHeaderV0For<Block, CBlock> {
    pub proof_of_election: ProofOfElection,
    pub receipt: ExecutionReceipt<<<CBlock as Block>::Header as Header>::Number, <CBlock as Block>::Hash, <<Block as Block>::Header as Header>::Number, <<Block as Block>::Header as Header>::Hash, u128>,
    pub estimated_bundle_weight: Weight,
    pub bundle_extrinsics_root: <<Block as Block>::Header as Header>::Hash,
}

Fields
proof_of_election: ProofOfElection

Proof of bundle producer election.

receipt: ExecutionReceipt<<<CBlock as Block>::Header as Header>::Number, <CBlock as Block>::Hash, <<Block as Block>::Header as Header>::Number, <<Block as Block>::Header as Header>::Hash, u128>

Execution receipt that should extend the receipt chain or add confirmations to the head receipt.

estimated_bundle_weight: Weight

The total (estimated) weight of all extrinsics in the bundle.

Used to prevent overloading the bundle with compute.

bundle_extrinsics_root: <<Block as Block>::Header as Header>::Hash

The Merkle root of all new extrinsics included in this bundle.