domain_client_operator
Module snap_sync

Structs

BlockImportingAcknowledgement

Notification with number of the block that is about to be imported and acknowledgement sender that pauses block production until the previous block is acknowledged.

ConsensusChainSyncParams

Provides parameters for domain snap sync synchronization with the consensus chain snap sync.

SnapSyncOrchestrator

Synchronizes consensus and domain chain snap sync.

SyncParams

Traits

LastDomainBlockReceiptProvider

Provides execution receipts for the last confirmed domain block.