domain_client_operator::snap_sync
Trait LastDomainBlockReceiptProvider

pub trait LastDomainBlockReceiptProvider<Block: BlockT, CBlock: BlockT>: Sync + Send {
    // Required method
    fn get_execution_receipt<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Option<ExecutionReceiptFor<Block::Header, CBlock, Balance>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Provides execution receipts for the last confirmed domain block.

Required Methods

fn get_execution_receipt<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Option<ExecutionReceiptFor<Block::Header, CBlock, Balance>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns execution receipts for the last confirmed domain block.